refactor codebase into components/hooks folder; add test files for Form and MessagesList components
