+++ client-app/src/App.spec.js
@@ -6,7 +6,7 @@ test.use({ viewport: { width: 500, height: 500 } });
 test('Verify components exist in App.js', async ({ mount }) => {
   const component = await mount(<App />);
   await expect(component).toContainText('Send a message');
-  await expect(component).toContainText('save to reload');
+  await expect(component).toContainText('Submit');
 });
 
 // goto url
