+++ client-app/src/components/Form.js
@@ -1,0 +1,33 @@
+import { useState } from "react";
+import axios from 'axios';
+
+export const Form = () => {
+  const [input, setInput] = useState("");
+
+  const handleChange = (e) => setInput(e.target.value);
+  const handleSubmit = () => {
+    axios.post(`${process.env.REACT_APP_API_GATEWAY_URI}/messages`, {
+      message: input
+    }).then(result => {
+      // do something with the response sent from server
+      if(result) {
+        const data = result.data;
+        console.log(data);
+      }
+      
+      // without websockets you'll have to refresh the page 
+      // to see 'messages' update with the new message
+    })
+    .catch(err => {
+      console.error(`Error when posting a message, ${err}`);
+    });
+  };
+
+  return (
+    <section>
+      <label> Send a message </label>
+      <input type="text" onChange={e => handleChange(e)} />
+      <button onClick={handleSubmit}> Submit </button>
+    </section>
+  );
+}
